requirements.txt
streamlit
replicate
python-dotenv
numpy
matplotlib
opencv-python
Pillow
scikit-learn
scipy